View Transaction Statement feature has been implemented and resolve some bugs. sprint3 is now completed with all features
