+++ .env
@@ -1,3 +1,3 @@
 DATABASE_URL=postgresql://postgres:1993@localhost:5432/payment_system_six
 
-JWT_SECRET = marianaConsultancy
+JWT_SECRET = HelloGo
